requirements.txt
django==2.0.4
python-dateutil==2.7.2